Added testimonials to database
