--- resume-website/src/_constants/Jobs.tsx
@@ -1,244 +1,0 @@
-import {IExperienceEntry} from "@/_components/typography/ExperienceEntry";
-
-export const jobData: IExperienceEntry[] = [
-    {
-        job: "Senior Software Engineer",
-        jobType: "Full-time",
-        company: "PKI Solutions",
-        startDate: new Date(2024, 3, 4),
-        summary: ["Developed a Next.js application from the ground up, integrating Single Sign-On (SSO) and CosmosDB for secure and scalable data management."
-            , "Led three technical projects from design to implementation, delivering impactful solutions directly to customers."
-            , "Established best coding practices and documentation, improving team efficiency and streamlining onboarding for new developers."
-            , "Quickly upskilled on a large monolithic ASP.NET Core application, contributing production-ready code within the first week."
-            , "Implemented modern C# methodologies on a legacy application, enhancing maintainability and performance."
-            , "Deployed client-side React pages on an ASP.NET Core website using Webpack, optimizing front-end performance and user experience."],
-        commuteType: "On-site",
-        location: "Portland, Oregon, United States", skills: [
-            "Project Coordination",
-            "ASP.NET Core",
-            "Azure DevOps Services",
-            "Technical Leadership",
-            "TypeScript",
-            "Technical Issues Resolution",
-            "Testing and Debugging",
-            "API Management",
-            "Microsoft Azure",
-            "Responsive UI/UX Design",
-            "Wireframing and Design",
-            "GitHub",
-            "Next.js",
-            "Express.js",
-            "Single Sign-On (SSO)",
-            "Entity Framework (EF) Core",
-            "Performance Optimization",
-            "Cross-functional Collaborations",
-            "Code Review",
-            "Software Development Life Cycle (SDLC)",
-            "Data Migration",
-            "Bitbucket",
-            "ASP.NET MVC",
-            "SAML 2.0",
-            "React.js",
-            "REST APIs",
-            "C#",
-            "New Technologies Integration",
-            "Internet Information Services (IIS)",
-            "Azure Cosmos DB",
-            "Amazon Web Services (AWS)",
-            "Database Management",
-        ]
-    },
-    {
-        job: "Cloud Software Engineer",
-        jobType: "Full-time",
-        company: "Intel Corporation",
-        startDate: new Date(2022, 8, 3),
-        finishDate: new Date(2024, 3, 4),
-        summary: ["Executed the launch of a website utilizing Bootstrap 5 by managing an initiative to enhance the website through the seamless migration from Bootstrap 5 to Material UI."
-            , "Conducted comprehensive research on existing Design Systems to disseminate findings and spearhead the migration and training initiatives for fellow developers, ensuring seamless adoption and integration."
-            , "Developed and executed an eBPF script by deploying and exposing it through a synergistic integration of AWS Lambda and AWS API Gateway."
-            , "Improved software quality by implementing the Jest unit test suite, conducting UIX testing with Selenium, and establishing Continuous Integration."
-            , "Elevated user experience and enhanced the application interface through strategic leadership in UI design meetings."],
-        commuteType: "Remote",
-        location: "Hillsboro, Oregon, United States",
-        skills: [
-            "Amazon Relational Database Service (RDS)",
-            "Project Coordination",
-            "Material-UI",
-            "Microsoft Excel",
-            "Technical Leadership",
-            "Technical Issues Resolution",
-            "Testing and Debugging",
-            "Python (Programming Language)",
-            "API Management",
-            "Responsive UI/UX Design",
-            "Wireframing and Design",
-            "NumPy",
-            "HTML",
-            "Performance Optimization",
-            "Linux",
-            "Cross-functional Collaborations",
-            "Code Review",
-            "Software Development Life Cycle (SDLC)",
-            "Web Applications",
-            "Software Development",
-            "React.js",
-            "Jest",
-            "Cascading Style Sheets (CSS)",
-            "Redux.js",
-            "Amazon Simple Notification Service (SNS)",
-            "New Technologies Integration",
-            "SQL",
-            "Amazon S3",
-            "JavaScript",
-            "Representational State Transfer (REST)",
-            "Pandas (Software)",
-            "AWS Lambda",
-            "Database Management",
-            "Agile Methodologies",
-        ]
-    },
-    {
-        job: "Full Stack Software Engineer",
-        jobType: "Full-time",
-        company: "Intel Corporation",
-        startDate: new Date(2021, 0, 1),
-        finishDate: new Date(2022, 8, 3),
-        summary: ["Optimized developer efficiency in database management by supporting the migration of legacy database from Elasticsearch to advanced MongoDB Atlas.",
-            "Enhanced UI/UX standards for intuitive user experience by revamping the Flask website through the integration of Bootstrap 4 and jQuery.",
-            "Maximized user satisfaction ratings by minimizing application database query times from 30-60 seconds to 0.5 seconds.",
-            "Delivered responsive applications by executing full-stack development and leveraging MongoDB Atlas, Flask, and Bootstrap 4 technologies.",
-            "Contributed toward the professional development of four interns by providing mentorship and constructive feedback on areas for improvement. Guided team in effective project execution by facilitating SCRUM and planning sprint meetings."
-        ],
-        commuteType: "Remote",
-        location: "Hillsboro, Oregon, United States",
-        skills: [
-            "Amazon Relational Database Service (RDS)",
-            "Project Coordination",
-            "Microsoft Excel",
-            "Technical Leadership",
-            "Technical Issues Resolution",
-            "Testing and Debugging",
-            "Python (Programming Language)",
-            "API Management",
-            "Responsive UI/UX Design",
-            "Wireframing and Design",
-            "NumPy",
-            "MongoDB",
-            "HTML",
-            "Microsoft SQL Server",
-            "Performance Optimization",
-            "Linux",
-            "Cross-functional Collaborations",
-            "Code Review",
-            "Software Development Life Cycle (SDLC)",
-            "Data Migration",
-            "Web Applications",
-            "Software Development",
-            "Jest",
-            "Cascading Style Sheets (CSS)",
-            "Amazon Simple Notification Service (SNS)",
-            "New Technologies Integration",
-            "SQL",
-            "Amazon S3",
-            "Amazon Web Services (AWS)",
-            "JavaScript",
-            "Representational State Transfer (REST)",
-            "Pandas (Software)",
-            "AWS Lambda",
-            "Database Management",
-            "Agile Methodologies",
-            "Elastic Stack (ELK)",
-        ]
-    },
-    {
-        job: "Cloud Automation Undergraduate Intern",
-        jobType: "Contract",
-        company: "Gridworks IC",
-        startDate: new Date(2020, 5, 14),
-        finishDate: new Date(2020, 11, 30),
-        summary: [
-            "Streamlined testing process by implementing Continuous Integration with PyTest, Selenium, and GitHub Action Runners.",
-            "Designed wireframes and suggested UI enhancements using Microsoft Visio, with a commitment to effective design solutions.",
-            "Restructured Flask website aesthetics through custom CSS, Bootstrap 4, and JavaScript for augmented visual appeal.",
-        ],
-        commuteType: "On-Site",
-        location: "Portland, Oregon, United States",
-        skills: [
-            "Elixir",
-            "Project Coordination",
-            "Microsoft Excel",
-            "Technical Leadership",
-            "Technical Issues Resolution",
-            "Testing and Debugging",
-            "Python (Programming Language)",
-            "API Management",
-            "Responsive UI/UX Design",
-            "Wireframing and Design",
-            "NumPy",
-            "HTML",
-            "Performance Optimization",
-            "Linux",
-            "Cross-functional Collaborations",
-            "Code Review",
-            "Software Development Life Cycle (SDLC)",
-            "Web Applications",
-            "Software Development",
-            "Cascading Style Sheets (CSS)",
-            "New Technologies Integration",
-            "SQL",
-            "Amazon Web Services (AWS)",
-            "JavaScript",
-            "Representational State Transfer (REST)",
-            "Pandas (Software)",
-            "Database Management",
-            "Agile Methodologies",
-            "Elastic Stack (ELK)",
-        ]
-    },
-    {
-        job: "Software Engineer Intern",
-        jobType: "Contract",
-        company: "Intel Corporation",
-        startDate: new Date(2019, 8, 2),
-        finishDate: new Date(2019, 11, 31),
-        summary: [
-            "Saved $100K+ monthly by utilizing a state-managed open-source mapping database to optimize customer routes and drive cost-effective operations.",
-            "Minimized the downtime for the data migration process of 640K records by incorporating asynchronous bulk SQL inserts and Django-based Python script.",
-            "Evaluated ERDs from two databases to decode complex data structures and understand intricate relationships in systems.",
-        ],
-        commuteType: "Remote",
-        location: "Hillsboro, Oregon, United States",
-        skills: [
-            "Elixir",
-            "Project Coordination",
-            "Microsoft Excel",
-            "Technical Leadership",
-            "Technical Issues Resolution",
-            "Testing and Debugging",
-            "Python (Programming Language)",
-            "Wireframing and Design",
-            "NumPy",
-            "HTML",
-            "Microsoft SQL Server",
-            "Performance Optimization",
-            "Linux",
-            "Cross-functional Collaborations",
-            "Code Review",
-            "Software Development Life Cycle (SDLC)",
-            "Data Migration",
-            "Web Applications",
-            "Software Development",
-            "React.js",
-            "Jest",
-            "New Technologies Integration",
-            "SQL",
-            "JavaScript",
-            "Representational State Transfer (REST)",
-            "Pandas (Software)",
-            "Database Management",
-            "Agile Methodologies",
-        ]
-    }
-
-];
