--- resume-website/src/_constants/Testimonials.tsx
@@ -1,56 +1,0 @@
-import {ITestimonialCard} from "@/_components/cards/TestimonialCard";
-
-type TTestimonialData = ITestimonialCard[]
-
-
-export const data: TTestimonialData = [
-    {
-        Recommender: "Robert Bosch",
-        externalUrl: "https://www.linkedin.com/in/rbosch1",
-        date: "April 2, 2025",
-        message: "Andrew Doser is a very adept software engineer who excels in navigating complex legacy systems and consistently delivers clean, concise, and thorough solutions. He has a knack for breaking down complex problems into simple solutions and considering valid edge cases that escape the consideration of other engineers he is working with, ensuring robust and stable systems. I greatly appreciated the wit, candor, and perspective Andrew brought to the team; skills that contributed to the success of many projects. I would highly recommend him for any senior software engineering role."
-    },
-    {
-        Recommender: "Alex Bishop",
-        externalUrl: "https://www.linkedin.com/in/alex-bishop-572430159/",
-        date: "April 1, 2025",
-        message: "I can say without a doubt Andrew is a results driven engineer. I have worked with him multiple times throughout my career and each time it has been an excellent and highly productive environment he facilitated. Andrew has a keen eye for details that he shows time and time again in code review as well as a multi-faceted and diverse knowledge base on software development, particularly with front-end development. Any team will no doubt be enriched with Andrew's dutiful work ethic and deep engineering knowledge. Most importantly of all, Andrew is a pleasant and sociable individual who can help encourage good teamwork and positive rapport with everyone he is involved with. Anyone who reads this should know that Andrew brings unreplaceable value to any projects he works on, with both his expertise and his personality."
-    },
-    {
-        Recommender: "Monica Ene-Pietrosanu",
-        externalUrl: "https://www.linkedin.com/in/monica-ene-pietrosanu",
-        date: "January 30, 2024",
-        message: "I know Andrew for 4+ years, when he joined my organization as a intern. Based on his great work we hired him as a full time software engineer one year later. He is a solid engineer, skilled in cloud software development and AWS technologies. Always ready to tackle difficult problems and to drive them to solution, Andrew was a great contributor in our SaaS products development team. His solid technical skills, constructive attitude, ability to ramp up fast on new projects/technologies as well as strong communication skills make him an great asset to any software engineering team."
-    },
-    {
-        Recommender: "David Klein",
-        externalUrl: "https://www.linkedin.com/in/dtklein7",
-        date: "January 29, 2024",
-        message: "Andrew is a driven engineer who excitedly tackles multifaceted problems. He takes joy understanding the nuances of technologies and making informed decisions about what tools are most effective for a given domain."
-    },
-    {
-        Recommender: "Radu Malliu",
-        externalUrl: "https://www.linkedin.com/in/radumalliu",
-        date: "January 25, 2024",
-        message: "I've worked with Andrew for 1.5 years. Throughout that time, I've consistently seen him deliver solutions that were reliable, efficient and well thought out. His focus on quality also shone through. I would recommend him to any team looking for a talented software engineer with a broad skill set."
-    },
-    {
-        Recommender: "Jassy Lee",
-        externalUrl: "https://www.linkedin.com/in/jassy-lee-a4a44b4/",
-        date: "January 23,2024",
-        message: "Andrew has been in my team for couple of years. He is a great developer with both front-end and back-end technologies. He has a strong foundation in languages such as Python, JavaScript and also database like MongoDB. He is able to seamlessly integrate these technologies to create cohesive systems. Andrew is also a good team player. He always makes an efficient conversation with team members with different backgrounds. His collaborative spirit and positive attitude is a pleasure to work with. I have no doubt Andrew would be a good asset to any development team."
-    },
-    {
-        Recommender: "David Maldonado Moreno",
-        externalUrl: "https://www.linkedin.com/in/david-maldonado-b5230370",
-        date: "January 23, 2024",
-        message: "I worked with Andrew on 2 different projects and I can say with complete certainty that he is an excellent engineer capable of solving complex problems and providing solutions very quickly and efficiently.\n" +
-            "I hope that, in the future, our paths will cross again and we can work together once more. Much success."
-    },
-    {
-        Recommender: "Anil Singh",
-        externalUrl: "https://www.linkedin.com/in/anilgavendrasingh",
-        date: "November 4, 2019",
-        message: "Andrew is a talented software engineer and team player. He is able to learn quickly and take on any task put in front of him. Well liked by the team and a great personality. "
-    }
-];
